change delete event to the admin folder
