+++ server/router/index.js
@@ -54,7 +54,6 @@ router.post(
 router.post('/cancelUser', validationCancelReg, cancelRegistration);
 router.post('/login', loginValidation, checkEmailIfExist, checkPassword, login);
 
-router.patch('/deleteevent', deleteEvent);
 
 // should be protected
 router.use(protectedRoute);
@@ -62,6 +61,7 @@ router.post('/event', validateEvent, createEvent);
 router.patch('/attendance', validateAttendence, checkUserBooking, signAttendance);
 router.get('/users', getUsersData);
 router.get('/event/:eventcode/users', getUsersEvent);
+router.patch('/deleteevent', deleteEvent);
 
 router.use(clientError);
 router.use(serverError);
